Fix issue where config only read once

+++ GreenhouseBuff/GreenhouseBuff/ModPatches/Beehive.cs
@@ -26,7 +26,7 @@ namespace GreenhouseBuff.ModPatches
                     (float)codes[i + 1].operand == 5f && codes[i + 2].opcode == OpCodes.Add)
                 {
                     // Replace it with temp += beeTempBonus
-                    codes[i + 1].operand = GreenhouseBuffConfig.Loaded.BeehiveTempMod;  // Change the operand to the loaded config value
+                    codes[i + 1].operand = (float)GreenhouseBuffConfig.Loaded.BeehiveTempMod;  // Change the operand to the loaded config value
                 }
             }
