+++ GreenhouseBuff/GreenhouseBuff/ModPatches/Farmland.cs
@@ -32,7 +32,7 @@ namespace GreenhouseBuff.ModPatches
                     (float)codes[i + 1].operand == 5f && codes[i + 2].opcode == OpCodes.Add)
                 {
                     // Replace it with conds.Temperature += farmlandTempBonus
-                    codes[i + 1].operand = GreenhouseBuffConfig.Loaded.FarmlandTempMod;  // Change the operand to the loaded config value
+                    codes[i + 1].operand = (float)GreenhouseBuffConfig.Loaded.FarmlandTempMod;  // Change the operand to the loaded config value
                 }
             }
